src/mcpo/mcp_servers/memory/requirements.txt
fastapi
uvicorn
pydantic
pydantic[email]
weaviate-client
numpy
requests
